% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT_tools.R
\name{.DTFitOptions}
\alias{.DTFitOptions}
\title{Fit default DT Options}
\usage{
.DTFitOptions(
  df,
  dom = getOption("InvenceRVisualTools.DT.dom"),
  buttons = getOption("InvenceRVisualTools.DT.buttons"),
  minPageLength = getOption("InvenceRVisualTools.DT.minPageLength"),
  hideColNames = NULL,
  useRegExp = FALSE,
  autoWidth = TRUE,
  columnDefs = NULL,
  JS.compact = TRUE
)
}
\arguments{
\item{df}{A data.frame to be shown}
}
\value{
A list of DT options.
}
\description{
Internal function used in DTSimple.
}
\examples{
.DTFitOptions(df)
}
