% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT_tools.R
\name{DTSimple}
\alias{DTSimple}
\title{Create an HTML table widget using the DataTables library in a simple way}
\usage{
DTSimple(
  df,
  dom = getOption("InvenceRVisualTools.DT.dom"),
  buttons = getOption("InvenceRVisualTools.DT.buttons"),
  minPageLength = getOption("DT.minPageLength"),
  extensions = "Buttons",
  opt_class = "display nowrap",
  selection = "single",
  filter = "none",
  rownames = FALSE,
  hideColNames = NULL,
  width = NULL,
  height = NULL,
  useRegExp = FALSE,
  editable = FALSE,
  autoWidth = TRUE,
  columnDefs = NULL,
  escape = FALSE,
  callback = JS("return table;"),
  options = list()
)
}
\arguments{
\item{df}{A data.frame to be shown}
}
\value{
An HTML table widget using the DataTables library.
}
\description{
Call to DT::datatable with usefull adapted defaults
}
\examples{
DTSimple(df)
}
