% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_tools.R
\name{PlotLyXts}
\alias{PlotLyXts}
\title{Creates a time series plotly chart}
\usage{
PlotLyXts(
  x,
  width = "100\%",
  type = "scatter",
  mode = "lines",
  connectgaps = TRUE,
  line_shape = NULL,
  verbose = FALSE
)
}
\arguments{
\item{x}{A xts object}

\item{with}{The width of the widget in accepted formats as "100\%" or "400px"}
}
\value{
A plotly object.
}
\description{
Creates a time series plotly chart from a xts object and
}
\examples{
PlotLySlider(p, 'D')
}
